Fix: Change max_pwm type from uint8_t to uint16_t in running modes to fit the actual range of PWM values and prevent overflow issues.
Fix: Add check for expected_size before reading buffer in serial_in to avoid incorrectly flushing the buffer when expected_size is zero, preventing potential data corruption.

+++ Core/state_machine/src/running_modes/running_pid.c
@@ -36,8 +36,8 @@ void running_pid(const StateMachine* const sm) {
 void running_pid_to_stopped(void) {
     const PidStruct* pid = get_pid();
 
-    const uint8_t max_pwm_save = pid->max_pwm;
-    uint8_t max_pwm = max_pwm_save;
+    const uint16_t max_pwm_save = pid->max_pwm;
+    uint16_t max_pwm = max_pwm_save;
 
     while (pid->max_pwm) {
         if (!update_pid()) continue;
